backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
psycopg[binary]==3.3.0
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.2
gunicorn


